Fixing up typedef getting rid of old uIntX replacing with uintX_t

+++ lib/objgfx40/objgfx40/objgfx40.h
@@ -79,7 +79,7 @@ class ogSurface {
 #else
  protected:
 #endif
-  float        version;
+  double        version;
   void       * buffer;
   ogSurface  * owner;
   uInt32     * lineOfs;
